Refactor file hierarchy
