--- src/components/sidebar/routePathView/newRoutePathView.scss
@@ -1,16 +1,0 @@
-@import '../../shared/styles/common';
-@import '../../shared/styles/flexbox';
-@import '../../shared/styles/form';
-
-.newRoutePathViewTab {
-    padding: 15px;
-    overflow-y: scroll;
-    width: $sidebarWide;
-
-    .staticInfo {
-        margin-top: 5px;
-        font-size: $smallFontSize;
-        margin-right: 10px;
-        font-weight: bold;
-    }
-}
