+++ src/components/sidebar/routePathView/routePathInfoTab/linkListView.scss
@@ -1,6 +1,6 @@
-@import '../../shared/styles/common';
-@import '../../shared/styles/flexbox';
-@import '../../shared/styles/form';
+@import '../../../shared/styles/common';
+@import '../../../shared/styles/flexbox';
+@import '../../../shared/styles/form';
 
 .linkListView {
