+++ src/components/sidebar/routePathView/routePathInfoTab/routePathInfoTab.scss
@@ -1,6 +1,6 @@
-@import '../../shared/styles/common';
-@import '../../shared/styles/flexbox';
-@import '../../shared/styles/form';
+@import '../../../shared/styles/common';
+@import '../../../shared/styles/flexbox';
+@import '../../../shared/styles/form';
 
 .routePathTab {
     padding: 15px 15px 45px 15px;
