+++ src/components/sidebar/routePathView/routePathView.scss
@@ -7,6 +7,10 @@
     width: $sidebarWide;
     height: 100%;
 
+    .content {
+        padding: 15px;
+    }
+
     .routePathViewHeader {
         display: flex;
     }
@@ -16,5 +20,12 @@
         font-weight: bold;
         color: #898989;
         font-size: 1rem;
+    }
+
+    .staticInfo {
+        margin-top: 5px;
+        font-size: $smallFontSize;
+        margin-right: 10px;
+        font-weight: bold;
     }
 }
